feat(qti/kodiak): add support for RB3Gen2 platform
RB3Gen2 is an IoT platform based on Kodiak SoC. Details about this
platform can be found here [1]. The boot flow with TF-A/OP-TEE is:

  PBL (ROM) -> XBL -> BL2 -> BL31 -> BL33 -> Normal world OS
                              |
                               --> BL32

Steps to build TF-A for RB3Gen2:

$ make -j`nproc` PLAT=rb3gen2 SPD=opteed QTISECLIB_PATH=<qtiseclib-path>
	BL32=<path-to-optee-bin> BL33=<path-to-os-bootloader-bin> fip all

$ ./tools/qti/generate_fip_elf.sh build/rb3gen2/release/fip.bin 0x9fc00000

The resulting fip.elf should be flashed in uefi_a partition on UFS flash
storage.

Note here that the bl2.elf generated requires to be signed as TZ image
with QTI signing involved. There is an ongoing effort to enable OEM only
signing for future boards support.

[1] https://www.qualcomm.com/developer/hardware/rb3-gen-2-development-kit

Change-Id: Ic19f28a5f559a0da28337f2d8da0d0e289a94514
Signed-off-by: Sumit Garg <sumit.garg@oss.qualcomm.com>